Separated operators.py file into multiple files, reworking the args mutation.
